Q173 Binary Search Tree Iterator
Implement an iterator over a binary search tree (BST). Your iterator
will be initialized with the root node of a BST.

Calling next() will return the next smallest number in the BST.

Note: next() and hasNext() should run in average O(1) time and uses O(h)
memory, where h is the height of the tree.